Add recipes
Got carried away and forgot to make smaller commits.
Added two recipes with description, ingredients and steps.
Added a hyperlink button on each recipe page to the homepage.
Created an images folder at the root to store recipe images.
On the homepage, put the recipe links into an unordered list.